+++ pytorch_lightning/plugins/legacy/ddp_sequential_plugin.py
@@ -21,7 +21,7 @@ from torch.nn.parallel import DistributedDataParallel
 
 from pytorch_lightning import _logger as log
 from pytorch_lightning import LightningModule
-from pytorch_lightning.plugins.rpc_plugin import RPCPlugin
+from pytorch_lightning.plugins.legacy.rpc_plugin import RPCPlugin
 from pytorch_lightning.utilities import _FAIRSCALE_PIPE_AVAILABLE, rank_zero_only
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
