+++ pytorch_lightning/plugins/legacy/native_amp.py
@@ -17,7 +17,7 @@ import torch
 from torch.optim import Optimizer
 
 from pytorch_lightning.core.optimizer import LightningOptimizer
-from pytorch_lightning.plugins.precision_plugin import PrecisionPlugin
+from pytorch_lightning.plugins.legacy.precision_plugin import PrecisionPlugin
 
 
 class NativeAMPPlugin(PrecisionPlugin):
