+++ pytorch_lightning/plugins/legacy/plugin.py
